requirements.txt
flask
google-genai
python-dotenv  # <-- Add this line
gunicorn